return a tibble with ancestral population of different segments

Uses link_ancestors from tskit, internally using python, to return a tibble with everything you need to know.

Usage

ancestral_segs(ts, focal_nodes_list, founder_nodes, nodes_tib, indiv_tib)

Arguments

ts

the tree sequence. You pass it as the object in py in the global env. i.e., as py$ts, for example.

focal_nodes_list

an unnamed list. Each element is a vector of integer node ids in which you want to track ancestry segments. It each vector must include all the focal individuals of each cohort.

founder_nodes

vector of integer ids of nodes that are considered the founders (typically the earliest ancestors of known population origin).

nodes_tib

tibble of the nodes. This is the nodes_tib component of the list returned by ts_nodes_and_inds().

indiv_tib

tibble of the indivs. This is the indiv_tib component of the list returned by ts_nodes_and_inds().