Add a column called cPOS that is the cumulative genome position

SLiM does everythign in terms of cumulative genome positions (i.e., there is just one chromosome that happens to have locations where the recombination fraction if 1/2). So, if we have a genome with coordinates of chrom, pos, we might want to get the cPOS for each. Doing so can be done efficiently by precomputing the cumulative start point of each chromosome like we have in the package data object mykiss_chroms.

Usage

cumulative_genome_position(X, Cumul)

Arguments

X

a tibble that has columns of chrom and pos

Cumul

a tibble that has, at a minimum, the columns chrom, and cumul_start. It will often be easiest to pass in your genome chromosome cumulative tibble like mykiss_chroms

Value

Returns a tibble like X but with an additional column cPOS that has the cumulative position.

Examples

cumulative_genome_position(example_spp_diag_markers, mykiss_chroms)
#> # A tibble: 1,297 × 4
#>    diag_spp chrom      pos     cPOS
#>       <int> <chr>    <int>    <dbl>
#>  1        1 omy01  5604476  5604476
#>  2        2 omy01  5604516  5604516
#>  3        3 omy01  5604546  5604546
#>  4        1 omy01  5604556  5604556
#>  5        2 omy01  8362156  8362156
#>  6        2 omy01  9582183  9582183
#>  7        1 omy01 12303335 12303335
#>  8        2 omy01 16142369 16142369
#>  9        2 omy01 16142395 16142395
#> 10        2 omy01 16206842 16206842
#> # ℹ 1,287 more rows