Place diagnostic marker alleles within individauls according to their ancestry haplotypes

Once individuals have been simulated and you have the individual segments within them, you can easily simulate the alleles of the species-diagnostic markers. That is what this function does.

Usage

place_diagnostic_markers(Markers, IndivSegs)

Arguments

Markers

a tibble that must have columns:

diag_spp_idx, which gives a 1-based index of the species for which the 1 allele is fixed at the marker (with all other species being fixed for the 0 allele).

cPOS, which are positions in the cumulative base pairs that SLiM uses.

chrom: the name of the chromosome (in the actual genome being modeled)

pos: the position in base pairs on chromosome chrom. Each row of this tibble must specify a unique species-diagnostic marker, and all of the species-diagnostic markers you wish to place into the simulated individuals should be included in this tibble.

IndivSegs

A tibble of individual segments like that which is returned by the function indiv_ancestry_tracts(). Though it can have arbitrary columns, it must include at least the following:

ped_id: The pedigree id of the individual

left: the left endpoint of the interval

right: the right endpoint of the interval

anc1: the 1-based index of the ancestry of haplotype 1 within the interval.

anc2: the 1-based index of the ancestry of haplotype 2 within the interval.

Value

This returns a tibble with the columns diag_spp_idx, chrom, pos, ped_id, and geno. This is almost what is needed to hand the data off to the ancestral_n_segments() function in the MixedUpParents package. You just have to change some column names and denote diag_spp by A, B, C.

Details

For this to work, you have to have both bedtools and awk on your PATH.

Examples

# Prepare Markers. Name the columns right, and get cumulative cPOS in there.
M <- example_spp_diag_markers %>%
  dplyr::filter(diag_spp < 3) %>%
  dplyr::rename(diag_spp_idx = diag_spp) %>%
  cumulative_genome_position(Cumul = mykiss_chroms)

# Now, also get some example indiv-segs. We have some as a package
# data object.
IS <- example_indiv_segs

# Then run the function:
place_diagnostic_markers(M, IS)
#> Running bedtools in /var/folders/ln/5brnn1wx0vj58fhlbbdjm8h40000gp/T//Rtmp1Hl8ZY/file13d4a5f18afa8
#> Rows: 10580 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (1): chrom
#> dbl (4): diag_spp_idx, pos, ped_id, geno
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 10,580 × 5
#>    diag_spp_idx chrom      pos ped_id  geno
#>           <dbl> <chr>    <dbl>  <dbl> <dbl>
#>  1            1 omy01  5604476  20055     2
#>  2            2 omy01  5604516  20055     0
#>  3            1 omy01  5604556  20055     2
#>  4            2 omy01  8362156  20055     0
#>  5            2 omy01  9582183  20055     0
#>  6            1 omy01 12303335  20055     2
#>  7            2 omy01 16142369  20055     0
#>  8            2 omy01 16142395  20055     0
#>  9            2 omy01 16206842  20055     0
#> 10            2 omy01 16206875  20055     0
#> # ℹ 10,570 more rows