Plot the ancestry copy number along segments in a group of individuals

More documentation needed. Basically just plots what comes out of indiv_ancestry_tracts(). Currently assumes just two ancestries. Plots individuals in order of admixture fraction.

Usage

plot_ancestry_tracts(IndSegs, chrom_ends = NULL)

Arguments

IndSegs

the tibble that comes out of indiv_ancestry_tracts().

chrom_ends

if desired, a tibble with column chrom giving the chromosome name and column end giving the cumulative base pair position of the end of the chromosome. Makes the plot return horizontal lines at the chromosome boundaries.

Value

returns a ggplot object