Simulate a complete data set to run in MixedUpParents

There is a lot that goes into this, and note that you have to have bedtools and bcftools installed on your system, etc. Also, it is currently up to the user to make sure that the parameters used here are correctly set for the SLiM file used.

Usage

slim_sim_a_dataset(
  AF,
  genome_info,
  founder_pop_nums,
  slim_file,
  years_list,
  marker_pop_time_list,
  diagnostic_markers,
  slim_seed = sample(1:1e+06, 1),
  focal_nodes = NULL,
  num_gen = 2
)

Arguments

AF

a tibble of allele frequencies with columns that must have chrom and then pos, and then as many columns as you wish, each one holding the frequency of the 1 allele in a population. The names of the columns give the names of the species/subspecies.

genome_info

a tibble of cumulative base pairs of different chromosomes in the genome giving the positions in AF. This is necessary for easily translating the positions into continuous coordinates for SLiM. Must have the columns chrom, start, end, cumul_start, and cumul_end. See mykiss_chroms for an example.

founder_pop_nums

a vector of the number of diploids in each founder population. This should be commensurate with the beginning population size in the SLiM simulation file.

slim_file

relative or absolute path to the .slim file to be used for the SLiM simulation. This must have been written so that it creates:

VCF output in the format "p" + subpop.id + "-" + sim.cycle + ".vcf"

tree sequence recorded output named "SLiM.trees"

years_list

A named list. Each name is an integer signifying a SLiM population, and the value is an integer vector of tree-seq years for which individuals/nodes in the tree seq from the named population should be retained in the tree-sequence. See the documentation for read_and_filter_trees() for more information. This must be concordant with the SLiM file you have used.

marker_pop_time_list

a list of pops and years from which to extract the genotypes from the VCF files. This is a named list with names being the names of the population as "p1", "p2", etc. and the years as the SLiM years (they get higher later in time) from which the samples are desired. For example: list(p1 = 9:11, p2 = 9:11).

diagnostic_markers

a tibble with columns diag_spp, chrom, and pos. The values in diag_spp must include the species codes used as column headers in AF.

slim_seed

the single integer seed to use for SLiM. Gets set automatically from the R RNG if not set.

focal_nodes

a tibble specifying the populations (base-1 indexed) and cohorts (base-0 tskit time—the most recent is 0) for which we want to determine ancestry vectors. If this is left NULL then this will be done for all populations in all non-founder generations.

num_gen

The number of generations back to trace the relationships between the individuals. By default this is 2, which will get you parent-offspring, full- and half siblings, grandparent/grandchild, aunts and uncles and first cousins.

Details

Note also that as currently configured this infers the founder generation' (in tskit time) by the time at which the smallest ped_id existed.

Examples

AF <- cyclone_creek_var_freqs
genome_info <- mykiss_chroms
founder_pop_nums <- c(1000, 1000)
slim_file <- system.file(
  "SLiM-models/2-pop-10-gens-vcf-initialize.slim",
  package = "MixedUpSlimSims"
)
slim_seed <- 55
years_list <- list(`1` = 0:10, `2` = 0:10)
marker_pop_time_list <- list(p1 = 9:11, p2 = 9:11)
diagnostic_markers <- wct_rbt_yct_diagnostic_markers

if (FALSE) { # \dontrun{
# this takes more time than CRAN likes, so don't run it on Check
simmed <- slim_sim_a_dataset(
  AF = AF,
  genome_info = genome_info,
  founder_pop_nums = founder_pop_nums,
  slim_file = slim_file,
  slim_seed = slim_seed,
  years_list = years_list,
  marker_pop_time_list = marker_pop_time_list,
  diagnostic_markers = diagnostic_markers
)

} # }