Read in multiple SLiM formatted VCF file

Read in multiple SLiM formatted VCF file

Usage

slim_vcfs2tib(
  Flist,
  Dir = ".",
  long = TRUE,
  parse_path = TRUE,
  drop_vcf_id = TRUE,
  drop_node_ids = TRUE
)

Arguments

Flist

a list of pops and years to get, like list(p1 = 9:11, p2 = 9:11).

Dir

directory the files are in. Defaults to current working directory.

long

Logical. If true (the default) this returns the long format version. Otherwise it returns the wide format.

parse_path

Logical. If TRUE, the file name is parsed to get the population and the slim time (i.e. the cycle or generation in SLiM). Default = TRUE.

drop_vcf_id

Logical. If TRUE, drop the vcf_id column. Default = TRUE. Has no effect if long = FALSE.

drop_node_ids

Logical. If TRUE, drop the node_id1 and node_id2 columns. Has no effect if long = FALSE.