return a tibble of the nodes and indviduals and meta data from a tree sequence

The way this works is a little non-standard. It writes the input, ts into the global environment under the name xx__TTKKSS where python can act upon it. Then a bunch of other stuff gets written into the python environment which is then pulled back into a tibble. At the end, we try to remove all the xx__ variables that we put into different global environments.

Usage

ts_nodes_and_inds(ts)

Arguments

ts

the tree sequence object as you would access it from the R environment, for example, py$ts or py$tree, etc.