Subsample SNPs or individuals and simulated genotyping error

Once you have simulated a SLIM dataset with slim_sim_a_dataset(), you can use this function to do the following:

Downsample to a consistent subsample of some number NV variable markers and/or some number ND diagnostic ones.

Create missing data at the remaining markers at a simple rate MV or MD across variable or diagnostic markers, respectively.

Randomly toss each individual in the sample from the sample at a single rate S across all individuals, pops and times.

Downsample the number of individuals to a specific number or fraction by population and time (not yet implemented).

Usage

tweak_slim_simmed_data(
  X,
  NV = NA,
  ND = NA,
  S = NA,
  MV = NA,
  MD = NA,
  EV = NA,
  ED = NA
)

Arguments

X

the return object of slim_sim_a_dataset().

NV

the number of variable markers to randomly downsample to, retaining the same markers across all individuals and completely discarding the rest. If NV is greater than the total number of variable markers a warning is issued and the full set of markers is returned.

ND

the number of diagnostic markers to randomly downsample to, retaining the same markers across all individuals and completely discarding the rest. If ND is greater than the total number of diagnostic markers a warning is issued and the full set of markers is returned.

S

the fraction of sampled individuals to be randomly retained. So, if you want to have sampled 40% of the population then you would set S = 0.4.

MV

the missing data data rate for variable markers. Each genotype in the variable-marker data set (after subsampling to NV markers) has an independent probability of MV of being missing applied.

MD

the missing data data rate for diagnostic markers. Each genotype in the variable-marker data set (after subsampling to NV markers) has an independent probability of MV of being missing applied.

EV

the probability that any non-missing variable marker genotype in an individual has a genotyping error. For a genotype that is simulated to have a genotyping error, if the true genotype is a 1 we choose either 0 or 1, each with probability of 1/2. If the true genotype is a 0 or a 2, the genotyping error results in a 1. (Genotyping error to the opposite homozygote does not occur).

ED

the probability that any non-missing diagnostic marker genotypes in an individual has a genotyping error. For a genotype that is simulated to have a genotyping error, if the true genotype is a 1 we choose either 0 or 1, each with probability of 1/2. If the true genotype is a 0 or a 2, the genotyping error results in a 1. (Genotyping error to the opposite homozygote does not occur).

Value

This returns a list with the downsampled and genotyping-errored versions of the data set.

Examples

# get a data set to work with
X <- slim_sim_dataset_example

tweaked <- tweak_slim_simmed_data(
  X,
  NV = 600,
  ND = 1000,
  S = 0.7,
  MV = 0.1,
  MD = 0.15,
  EV = 0.01,
  ED = 0.005
)